fix: fix get date and display
