+++ src/actions/quizResult.ts
@@ -2,11 +2,11 @@
 import { auth } from '@clerk/nextjs/server';
 import { neon } from '@neondatabase/serverless';
 import { TQuiz_data } from '@/lib/definitions';
-import { getDate } from '@/lib/getDate';
 
 export async function createRecord(quiz_data: TQuiz_data[]) {
   const { userId } = auth();
-  const date = getDate();
+  const now = new Date();
+  const date = now.toISOString();
   try {
     if (!process.env.DATABASE_URL) {
       throw new Error('DATABASE_URL is not defined');
